requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Werkzeug==2.0.1
mysql-connector-python==8.0.26
SQLAlchemy==1.4.23
Jinja2==3.0.1
itsdangerous==2.0.1
click==8.0.1
python-dotenv==1.0.0 